+++ common/src/main/java/net/elemental_wizards_rpg/datagen/ElementalSmithingRecipeProvider.java
@@ -1,6 +1,7 @@
 package net.elemental_wizards_rpg.datagen;
 
 import net.elemental_wizards_rpg.item.armor.Armors;
+import net.elemental_wizards_rpg.item.weapons.WeaponsRegister;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.minecraft.item.Items;
 import net.minecraft.util.Identifier;
@@ -47,6 +48,50 @@ public class ElementalSmithingRecipeProvider extends SmithingRecipeGenerator {
                 Items.NETHERITE_INGOT,
                 Armors.netheriteWindArmor.armorSet()
         );
+        // Netherite Weapons
+        createSimpleSmithingRecipe(
+                "smithing_netherite_aqua_wand",
+                WeaponsRegister.aquaWand.item(),
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE,
+                Items.NETHERITE_INGOT,
+                WeaponsRegister.netheriteAquaWand.item()
+        );
+        createSimpleSmithingRecipe(
+                "smithing_netherite_terra_wand",
+                WeaponsRegister.terraWand.item(),
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE,
+                Items.NETHERITE_INGOT,
+                WeaponsRegister.netheriteTerraWand.item()
+        );
+        createSimpleSmithingRecipe(
+                "smithing_netherite_wind_wand",
+                WeaponsRegister.windWand.item(),
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE,
+                Items.NETHERITE_INGOT,
+                WeaponsRegister.netheriteWindWand.item()
+        );
+        createSimpleSmithingRecipe(
+                "smithing_netherite_aqua_staff",
+                WeaponsRegister.aquaStaff.item(),
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE,
+                Items.NETHERITE_INGOT,
+                WeaponsRegister.netheriteAquaStaff.item()
+        );
+        createSimpleSmithingRecipe(
+                "smithing_netherite_terra_staff",
+                WeaponsRegister.terraStaff.item(),
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE,
+                Items.NETHERITE_INGOT,
+                WeaponsRegister.netheriteTerraStaff.item()
+        );
+        createSimpleSmithingRecipe(
+                "smithing_netherite_wind_staff",
+                WeaponsRegister.windStaff.item(),
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE,
+                Items.NETHERITE_INGOT,
+                WeaponsRegister.netheriteWindStaff.item()
+        );
+
 
         // ==========================================
         // ARMORY RPG UPGRADES (With mod load conditions)