+++ fabric/src/main/java/com/elemental_wizards/fabric/client/FabricClient.java
@@ -7,5 +7,6 @@ public final class FabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         ElementalClient.init();
+        ElementalClient.registerParticleAppearances();
     }
 }